requirements.txt
Django==4.0.7
djangorestframework==3.13.1
option==2.1.0
psycopg2==2.9.3
python-dotenv==0.19.2
